Control_Uno - HF patch 1
1st iteration of hotfire control code for the uno. 

This file should provide basic functionality and be bug free. There could be some bugs and logic issues when we perform an integrated test. The code could also be cleaner.

The uno will handle logic and transmit commands to the mega. The mega will receive these commands and double check the logic, before executing the commands on the stand.